rustdoc: factor `document_type_layout` into its own module
